+++ site/src/components/MainNav.tsx
@@ -2,9 +2,9 @@
  * This file is the Gatsby-specific header. The Storybook header is
  * located storybook/header-preset/header
  */
-import { Link as NavLink, NavigationBar } from "@kaizen/component-library"
 import { withPrefix } from "gatsby"
 import * as React from "react"
+import { Link as NavLink, NavigationBar } from "./NavigationBar"
 
 type MainNavProps = {
   currentPath?: string