+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.60](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.3.59...@kaizen/draft-guidance-block@3.3.60) (2022-01-05)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [3.3.59](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.3.58...@kaizen/draft-guidance-block@3.3.59) (2022-01-04)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [3.3.58](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.3.57...@kaizen/draft-guidance-block@3.3.58) (2021-12-24)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [3.3.57](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.3.56...@kaizen/draft-guidance-block@3.3.57) (2021-12-22)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [3.3.56](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.3.55...@kaizen/draft-guidance-block@3.3.56) (2021-12-20)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block